--- a/presto-client/src/main/java/com/facebook/presto/client/FixJsonDataUtils.java
+++ b/presto-client/src/main/java/com/facebook/presto/client/FixJsonDataUtils.java
@@ -127,7 +127,7 @@ final class FixJsonDataUtils
         if (signature.isVarcharEnum()) {
             return String.class.cast(value);
         }
-        if (signature.isLongEnum()) {
+        if (signature.isBigintEnum()) {
             if (value instanceof String) {
                 return Long.parseLong((String) value);
             }