--- a/presto-common/src/main/java/com/facebook/presto/common/type/LongEnumParametricType.java
+++ b/presto-common/src/main/java/com/facebook/presto/common/type/LongEnumParametricType.java
@@ -13,9 +13,6 @@
  */
 package com.facebook.presto.common.type;
 
-import com.facebook.presto.common.QualifiedObjectName;
-import com.facebook.presto.common.type.LongEnumType.LongEnumMap;
-
 import java.util.List;
 
 import static java.lang.String.format;
@@ -23,44 +20,25 @@ import static java.lang.String.format;
 public final class LongEnumParametricType
         implements ParametricType
 {
-    private final QualifiedObjectName name;
-    private final LongEnumMap enumMap;
+    public static final LongEnumParametricType LONG_ENUM = new LongEnumParametricType();
 
-    public LongEnumParametricType(QualifiedObjectName name, LongEnumMap enumMap)
-    {
-        this.name = name;
-        this.enumMap = enumMap;
-    }
+    private LongEnumParametricType() {}
 
     @Override
     public String getName()
     {
-        return name.toString();
-    }
-
-    @Override
-    public TypeSignatureBase getTypeSignatureBase()
-    {
-        return TypeSignatureBase.of(name);
+        return StandardTypes.LONG_ENUM;
     }
 
     @Override
     public Type createType(List<TypeParameter> parameters)
     {
-        if (parameters.isEmpty()) {
-            return new LongEnumType(name, enumMap);
-        }
         checkArgument(parameters.size() == 1, "Enum type expects exactly one parameter, got %s", parameters);
         checkArgument(
                 parameters.get(0).getKind() == ParameterKind.LONG_ENUM,
                 "Enum definition expected, got %s",
                 parameters);
-        return new LongEnumType(name, parameters.get(0).getLongEnumMap());
-    }
-
-    public LongEnumMap getEnumMap()
-    {
-        return enumMap;
+        return new LongEnumType(parameters.get(0).getLongEnumMap());
     }
 
     private static void checkArgument(boolean argument, String format, Object... args)