--- a/presto-common/src/main/java/com/facebook/presto/common/type/LongEnumType.java
+++ b/presto-common/src/main/java/com/facebook/presto/common/type/LongEnumType.java
@@ -13,7 +13,6 @@
  */
 package com.facebook.presto.common.type;
 
-import com.facebook.presto.common.QualifiedObjectName;
 import com.facebook.presto.common.block.Block;
 import com.facebook.presto.common.function.SqlFunctionProperties;
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -29,6 +28,8 @@ import static com.facebook.presto.common.type.BigintType.BIGINT;
 import static com.facebook.presto.common.type.TypeUtils.normalizeEnumMap;
 import static com.facebook.presto.common.type.TypeUtils.validateEnumMap;
 import static java.lang.String.format;
+import static java.util.Locale.ENGLISH;
+import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toMap;
 
 public class LongEnumType
@@ -37,9 +38,9 @@ public class LongEnumType
 {
     private final LongEnumMap enumMap;
 
-    public LongEnumType(QualifiedObjectName name, LongEnumMap enumMap)
+    public LongEnumType(LongEnumMap enumMap)
     {
-        super(new TypeSignature(name, TypeSignatureParameter.of(enumMap)));
+        super(new TypeSignature(StandardTypes.LONG_ENUM, TypeSignatureParameter.of(enumMap)));
         this.enumMap = enumMap;
     }
 
@@ -74,21 +75,29 @@ public class LongEnumType
     @Override
     public String getDisplayName()
     {
-        return getTypeSignature().getBase();
+        return enumMap.getTypeName();
     }
 
     public static class LongEnumMap
     {
+        private final String typeName;
         private final Map<String, Long> enumMap;
         private final Map<Long, String> flippedEnumMap;
 
         @JsonCreator
-        public LongEnumMap(@JsonProperty("enumMap") Map<String, Long> enumMap)
+        public LongEnumMap(@JsonProperty("typeName") String typeName, @JsonProperty("enumMap") Map<String, Long> enumMap)
         {
-            validateEnumMap(enumMap);
+            validateEnumMap(requireNonNull(enumMap, "enumMap is null"));
+            this.typeName = requireNonNull(typeName.toLowerCase(ENGLISH), "typeName is null");
             this.enumMap = normalizeEnumMap(enumMap);
             this.flippedEnumMap = this.enumMap.entrySet().stream()
                     .collect(toMap(Map.Entry::getValue, Map.Entry::getKey));
+        }
+
+        @JsonProperty
+        public String getTypeName()
+        {
+            return typeName;
         }
 
         @JsonProperty
@@ -114,25 +123,22 @@ public class LongEnumType
 
             LongEnumMap other = (LongEnumMap) o;
 
-            return Objects.equals(this.enumMap, other.enumMap);
+            return Objects.equals(typeName, other.typeName) && Objects.equals(enumMap, other.enumMap);
         }
 
         @Override
         public String toString()
         {
-            return "enum:bigint{"
-                    + enumMap.entrySet()
-                    .stream()
+            return format("%s{%s}", typeName, enumMap.entrySet().stream()
                     .sorted(Comparator.comparing(Map.Entry::getKey))
                     .map(e -> format("\"%s\": %d", e.getKey().replaceAll("\"", "\"\""), e.getValue()))
-                    .collect(Collectors.joining(", "))
-                    + "}";
+                    .collect(Collectors.joining(", ")));
         }
 
         @Override
         public int hashCode()
         {
-            return Objects.hash(enumMap);
+            return Objects.hash(typeName, enumMap);
         }
     }