--- a/presto-common/src/main/java/com/facebook/presto/common/type/StandardTypes.java
+++ b/presto-common/src/main/java/com/facebook/presto/common/type/StandardTypes.java
@@ -45,6 +45,8 @@ public final class StandardTypes
     public static final String IPPREFIX = "ipprefix";
     public static final String GEOMETRY = "Geometry";
     public static final String BING_TILE = "BingTile";
+    public static final String LONG_ENUM = "LongEnum";
+    public static final String VARCHAR_ENUM = "VarcharEnum";
 
     private StandardTypes() {}
 }