… presto-common/src/main/java/com/facebook/presto/common/type/VarcharEnumType.java
… presto-common/src/test/java/com/facebook/presto/common/type/BenchmarkTypeSignatureParsing.java
… presto-common/src/test/java/com/facebook/presto/common/type/TestTypeSignature.java
… presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/AbstractSqlInvokedFunctionNamespaceManager.java
… presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/mysql/EnumTypeRowMapper.java
… presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/mysql/FunctionNamespaceDao.java
… presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/mysql/MySqlFunctionNamespaceManager.java
… presto-function-namespace-managers/src/main/java/com/facebook/presto/functionNamespace/testing/InMemoryFunctionNamespaceManager.java
… presto-main/src/main/java/com/facebook/presto/metadata/BuiltInTypeAndFunctionNamespaceManager.java
… presto-main/src/main/java/com/facebook/presto/metadata/FunctionAndTypeManager.java
… presto-main/src/main/java/com/facebook/presto/sql/analyzer/ExpressionTreeUtils.java
… presto-main/src/main/java/com/facebook/presto/type/EnumCasts.java
… presto-main/src/main/java/com/facebook/presto/type/LongEnumOperators.java
… presto-main/src/main/java/com/facebook/presto/util/JsonUtil.java
… presto-spi/src/main/java/com/facebook/presto/spi/function/FunctionNamespaceManager.java
… presto-tests/src/main/java/com/facebook/presto/tests/TestingPrestoClient.java
… presto-tests/src/test/java/com/facebook/presto/tests/TestEnums.java