Depended by https://github.com/facebookexternal/presto-facebook/pull/1372

Introduce the concept of UserDefinedType as spedified in spec16. User defined
types are managed by FunctionNamespaceManager.

The spec has more metadata specified for user defined types but we currently
only add
```
<user-defined type definition> ::= CREATE TYPE <user-defined type body>
<user-defined type body> ::= <schema-resolved user-defined type name>
    [ AS <representation> ]
<representation> ::=
    <predefined type>
  | <collection type>
  | <member list>
```
For <representation>, we use TypeSignature to represent all options above.

Enum type signature format is changed from `typename(enum:type{enum_map})` to
`enumtype(typename{enum_map})` so enum types can behave the same way as other
builtin parametric types that's fully managed by
BuiltInTypeAndFunctionNamespaceManager.

Test plan - Travis

```
== NO RELEASE NOTE ==
```