@rongrong 2020-12-30
There seems to be a slight performance regression for type signature parsing that I can't get to the bottom of:

Before the type signature format change:
```
Benchmark                                                       Mode  Cnt    Score   Error  Units
BenchmarkTypeSignatureParsing.parseBigLongEnumTypeSignature     avgt   20   60.107 ± 6.650  ms/op
BenchmarkTypeSignatureParsing.parseBigVarcharEnumTypeSignature  avgt   20  140.270 ± 4.333  ms/op
BenchmarkTypeSignatureParsing.parseRowTypeSignature             avgt   20    2.101 ± 0.196  ms/op
BenchmarkTypeSignatureParsing.parseRowTypeSignatureWithEnums    avgt   20    2.584 ± 0.473  ms/op
```
After:
```
Benchmark                                                       Mode  Cnt    Score   Error  Units
BenchmarkTypeSignatureParsing.parseBigLongEnumTypeSignature     avgt   20   63.391 ± 4.711  ms/op
BenchmarkTypeSignatureParsing.parseBigVarcharEnumTypeSignature  avgt   20  146.647 ± 7.203  ms/op
BenchmarkTypeSignatureParsing.parseRowTypeSignature             avgt   20    3.038 ± 0.331  ms/op
BenchmarkTypeSignatureParsing.parseRowTypeSignatureWithEnums    avgt   20    3.193 ± 0.356  ms/op
```