@rongrong 2021-01-14
> Can we split it into two PRs (not just commits) - one for UserDefined type that's of more general interest to the community and enum separately.

I tried. The problem is either the commit / PR that introduces user defined type is not really used (only a simple container class, and dead code), or I need to change enum to use it, which requires the refactor of enum.